Add files via upload
Use-case (Melbourne House Price Analysis)
==========================================
Your end goal is to create a model that can predict the price of house based on Car Parking availability, Landsize, BuildingArea, YearBuilt. For that we need to perform preprocessing activities to make our data compatible for the model

Perform the following tasks using SKLEARN and PANDAS:

1. Loading the data (pandas)
2. Seperate the data with features and label.
3. Check for missing data and repair the same
4. Check for categorical data and repair the same
5. Check for ordinal data and repair the same 
6. Perform Feature Scaling
7. 


You need to perform this assignment in two mindsets
1. Stats perspective
2. Domain Perspective

The usual  car parking alloted to each house as per govt policy is 1 NaN
The Minimum build area for the building as per govt norms is 80 NaN
Maximums buildings as per the census survey was built in year 2000. NaN