Add files via upload
House Loan Data Analysis
Project 1 
DESCRIPTION
For safe and secure lending experience, it's important to analyse the past data. In this project, you have to build a deep learning model to predict the chance of default for future loans using the historical data. As you will see, this dataset is highly imbalanced and includes a lot of features that make this problem more challenging.

Objective: Create a model that predicts whether or not an applicant will be able to repay a loan using historical data.
Domain: Finance
Analysis to be done: Perform data pre-processing and build a deep learning prediction model.
Steps to be done: 
⦁    Load the dataset that is given to you
⦁    Check for null values in the dataset
⦁    Print percentage of default to payer of the dataset for the TARGET column
⦁    Balance the dataset if the data is imbalanced
⦁    Plot the balanced data or imbalanced data
⦁    Encode the columns that is required for the model
⦁    Calculate Sensitivity as a metric
⦁    Calculate area under receiver operating characteristics curve

Following Tasks have been completed as part of this project.
1.	Verified the Null columns and %age of Null values in the data
2.	Filled in Null values for  Numerical features with MEAN
3.	Filled in Null values for  categorical features with MODE
4.	Verified the Feature collinearity using heat map 
5.	Used pair plots for various feature and Target combinations to know the  data separability/relationship
6.	Feature Engineering:
a.	Selected top 5 Categorical features using Chi-square Test
b.	Encoded the columns that is required for the model
c.	Applied PCA as well as Kernel  PCA to for feature extraction
7.	Applied Logistics Regression and trained the model
8.	Calculated Precision and Recall( Sensitivity)
9.	Calculated area under receiver operating characteristics curve
10.	Created Deep Learning model with below architecture
a.	Input layers with 80 Features
b.	Applied Batch normalization at Input as well as hidden layer
c.	4 hidden layers and Output layer with sigmoid activation function