Add files via upload
In this competition, we will be predicting the probability [0, 1] of a binary target column.
The data contains binary features (bin_*), nominal features (nom_*), ordinal features (ord_*) as well as (potentially cyclical) day (of the week) and month features. 
Since the purpose of this competition is to explore various encoding strategies, the data has been simplified in that there are no missing values.